Add support for arbitrarily nested source files
Supporting nested directories by replacing directory separators with
underscores is smelly, especially because the inverse operation is
needed as well (to go from underscores to directory separators). It's
not uncommon to have underscores in file names to begin with, which
leads to obvious breakage. The new solution is just to reflect the
source directory structure in the object directory (i.e.
src/myfolder/myfile.c -> bin/myfolder/myfile.o), which should be
impossible to mess up.

Since we're now going to arbitrary nesting, src/Makefile is no longer
required.