Vision Sensor Signature Filtering/Coordinate Translation and Addition of V5 Device Struct Padding
Summary:
Reworked the Vision Sensor API according to T524.

This added the functions `vision_get_by_sig` and `vision_read_by_sig`, which filter the incoming data by signature and should be quite useful. These take in 1 aligned sig ids as opposed to the default 0 aligned ids.

`vision_read_signature` and `vision_write_signature` are no longer exposed to the user because they are not tied to the Vision Utility; it is recommended to use the Vision Utility for setting signatures instead.

Additionally, `vision_read_signature` now returns a signature object instead of taking a pointer to a signature object.

Functionality for changing the (0,0) coordinate point for the field of view was also added, so objects' coordinates can now be expressed relative to the center of the field of view as opposed to the top left.

To facilitate this change, additional padding was added to the V5 Device Struct for storing data like this for each smart port. This change allowed for a reworking of the way that ADI data was stored. The padding size is currently determined by the ADI's needs, if there ever becomes a reason for more ADI data to be stored (which is unlikely) then this padding size will need to be increased, but this will have no effect on the Vision sensor's use of the padding.

Closes T518

Test Plan:
[x] Compiles
[x] Verify Symbols
[x] Test FOV changes with the Vision Sensor
```
lang=c
vision_set_zero_point(1, E_VISION_ZERO_CENTER);
while (true) {
  vision_object_s_t rtn = vision_get_by_size(1, 0);
  lcd_print(1, "left: %d, top: %d", rtn.left_coord, rtn.top_coord);
  delay(2);
}
```

[x] Test `get_by_sig` function
```
lang=c
while (true) {
  vision_object_s_t rtn = vision_get_by_sig(1, 0, 1);
  lcd_print(1, "sig: %d", rtn.signature);
  delay(2);
}
```
[x] Test ADI changes
```
adi_encoder_t enc = adi_encoder_init(1,2,true); //alternate with false
  while (true) {
    lcd_print(4, "%d", adi_encoder_get(enc));
    delay(50);
  }
```

Reviewers: O1 The Ori, berman5

Reviewed By: O1 The Ori, berman5

Subscribers: berman5

Tags: #zorp

Maniphest Tasks: T518, T524

Differential Revision: http://phabricator.purduesigbots.com/D80