✨ Add support for Generic Serial Communication
Summary:
Ties Generic Serial Comms to newlib, so reading/writing to a generic smart port can be treated like a file.

Closes T573, and as a result closes T309 and T305

Test Plan:
[x] Compiles
[x] Test general reading:
```
FILE* read = fopen("/dev/11", "rw");
char buf[50];
while(1) {
  printf("val: %d\n", fread(buf, 1, sizeof(buf), read));
  delay(20);
}
```
The above code prints just 0 if there is not a device on port 11, or a nonzero value if there is (I saw 50 with the motor not moving).

Note: The above test code required a slight modification to the kernel code where the checks for the device type in the registry were ignored, so I could pull generic data off of a Smart Motor. The registry checks were added back so that this code will only be used with actually generic devices in the future.

Reviewers: O1 The Ori, brookea

Reviewed By: O1 The Ori, brookea

Subscribers: brookea

Tags: #zorp

Maniphest Tasks: T573, T309, T305

Differential Revision: https://phabricator.purduesigbots.com/D146