🐛 Vision Sensor Fixes
Summary:
The Vision Sensor Utility has a range of signature IDs of 1-7, not 1-8 as we had before. Additionally, we had been assuming that the signatures are 0-aligned in VexOS, when they are actually 1-aligned (Setting SIG_1 would register as sig_id 2 in PROS).

Additionally, I discovered while testing the above thing that the objects in an object array given to `read_by_size()` or `read_by_sig()` are not changed when there are no objects found, so there's no way to tell when an object has left the FOV. This has been resolved by setting the object signatures to 255 (`VISION_OBJECT_ERR_SIG`).

Closes T606.

Test Plan:
Matches VCS range and alignment.

```
pros::Vision vis (20);
while(true) {
  pros::vision_object_s_t obj = vis.get_by_sig(0,1);
  printf("sig: %d\n", obj.signature);
  delay(20);
}
```

prints 1  when an object set for SIG_1 is found, 255 otherwise.

This also worked in the same manner with the `get_by_size()`, `read_by_size()`, and `read_by_sig()` functions.

Reviewers: O1 The Ori, berman5, brookea

Reviewed By: O1 The Ori, berman5, brookea

Subscribers: brookea

Tags: #zorp

Maniphest Tasks: T606

Differential Revision: https://phabricator.purduesigbots.com/D154