🐛 Fix conflicts with okapilib inclusion
Summary:
Closes T616 and T618.

Original idea was to make the motor ctors implicit, but that conflicts with the overloaded `=` operator. Left the motor ctors alone, and removed the use of literals in the example.

Also did some polishing of the comments, looks like I missed main.h when I prepped for 3.1.0

Test Plan:
[x] Compiles
[ ] Maintains functionality

Reviewers: O1 The Ori, berman5

Reviewed By: O1 The Ori, berman5

Tags: #zorp

Maniphest Tasks: T618, T616

Differential Revision: https://phabricator.purduesigbots.com/D176