🚸 Stop overflow with `motor_move()`
Summary:
I noticed when testing out clawbot code that using arcade drive with `motor_move()` caused undefined behavior when
the sum of the two controller inputs exceeded 128. This can be fixed easily in user code by clamping
the sum of the controller values to +-127, but is an extra bit of code that most novice users probably won't think
to do. That, and the fact that the bug isn't very clear (the bot just kinda stutters while driving, or turns the wrong
direction).

Test Plan:
[x] Test arcade control code with this change
```
void opcontrol() {
	pros::Controller master(pros::E_CONTROLLER_MASTER);
	pros::Motor left_mtr(11);
	pros::Motor right_mtr(10, true);
	while (true) {
		pros::lcd::print(0, "%d %d %d", (pros::lcd::read_buttons() & LCD_BTN_LEFT) >> 2,
		                 (pros::lcd::read_buttons() & LCD_BTN_CENTER) >> 1,
		                 (pros::lcd::read_buttons() & LCD_BTN_RIGHT) >> 0);
		int pwr = master.get_analog(ANALOG_LEFT_Y);
		int trn = master.get_analog(ANALOG_RIGHT_X);
		int left = pwr + trn;
		int right = pwr - trn;

		left_mtr = left;
		right_mtr = right;
		pros::delay(20);
	}
}
```
This code no longer has the issue of jittering or skipping.

Reviewers: O1 The Ori, berman5

Reviewed By: O1 The Ori, berman5

Tags: #zorp

Differential Revision: https://phabricator.purduesigbots.com/D189