✨Gyro Driver
Summary:
It was technically possible to use the gyro before, but now there's a full driver for it with similar functionality to PROS 2, in both C and C++.

Closes T622.

Test Plan:
Ran the following test code:
```
 void opcontrol() {
 	pros::Controller master(pros::E_CONTROLLER_MASTER);
 	pros::Motor left_mtr(1);
 	pros::Motor right_mtr(2);
	pros::motor_pid_s_t pid = pros::Motor::convert_pid(1.0, 0.0001, 1.0, 2.0);
	pros::lcd::print(2, "%d %d %d %d", pid.kp, pid.ki, pid.kd, pid.kf);
	left_mtr.set_vel_pid(pid);
 	pros::ADIGyro gyro ('c', 1/1.8);
	pros::motor_pid_full_s_t pidf = left_mtr.get_vel_pid();
	pros::lcd::print(3, "%d %d %d %d", pidf.kp, pidf.ki, pidf.kd, pidf.kf);
	// pros::motor_pid_s_t pid = pros::Motor::convert_pid(1.0, 0.0001, 1.0, 2.0);
	// left_mtr.set_vel_pid(pid);
 	uint32_t now = pros::millis();
 	while (true) {
 		pros::lcd::print(0, "%d %d %d", (pros::lcd::read_buttons() & LCD_BTN_LEFT) >> 2,
 		                 (pros::lcd::read_buttons() & LCD_BTN_CENTER) >> 1,
 		                 (pros::lcd::read_buttons() & LCD_BTN_RIGHT) >> 0);
 		pros::lcd::print(1, "%f", gyro.get_value());
 		int left = master.get_analog(ANALOG_LEFT_Y);
 		int right = master.get_analog(ANALOG_RIGHT_Y);

 		if (pros::millis() - now > 5000) {
 			gyro.reset();
 			now = pros::millis();
 		}

 		left_mtr = left;
 		right_mtr = right;
 		pros::delay(20);
 	}
 }
```

The gyro read 1800 for a half rotation with the first constructor, and 1000 for the second. In both cases, the gyro was reset every 5s, and held its position well when sitting still and resetting like this.

The gyro calibrates every time that the port is configured as a gyro, so I tested the check for calibration by defining two gyro objects, one in `initialize()` and the other in `opcontrol()`.

```
void initialize() {
	pros::lcd::initialize();
	pros::lcd::set_text(1, "Hello PROS User!");

	pros::lcd::register_btn1_cb(on_center_button);
	pros::ADIGyro gyro (3);
}
```
```
Same opcontrol code as previous example
```

This removed the awkward second of giant negative values returned from the gyro before the calibration check, and instead returned `0` (as expected) at the first read in op control.

The motor pid calls are to ensure that the gyro data isn't overwritten by devices on smart ports like before - a device on ADI port 1 would have been using the same data as the device on Smart Port 1. The output on the screen was "0 16 32 16" for both the call before and after the gyro initialization (and corresponding setting of the multiplier). This matches with the data that was sent to the motor (although it's out of order in the setting/getting, my bad). The gyro's calibration still worked as expected; the gyro returned 1000 for a rotation of 180 degrees.

Also tested encoder reversal and digital new press with the new button switch, which also worked as expected . The full test code for these tests and the gyro can be found in tests/adi.cpp.

Reviewers: O1 The Ori, berman5

Reviewed By: O1 The Ori, berman5

Subscribers: brookea, berman5

Tags: #zorp

Maniphest Tasks: T622

Differential Revision: https://phabricator.purduesigbots.com/D184