Move vdml_initialize into init_array
Summary:
Discovered a bug where it was possible for user to create tasks during global initialization and immediately attempt to claim VDML mutexes before they were initialized causing a segmentation fault. This change does the following:
 - Move mutex initialization into the init_array so they are created before any global constructors are invoked
 - The XXX init delay holds port mutexes so that other start-up tasks cannot access VDML why VEXos is sorting itself out

We also determined that `vdml_initialize` did not need to occur with the scheduler started.

Test Plan:
[x] Run a motor in a task that starts in `opcontrol()`
```
void motor_task(void* ign) {
	pros::Motor test_mtr (20);
	while (true) {
		test_mtr.move_velocity(100);
		pros::delay(5);
	}
}
void opcontrol() {
	pros::Task test_task (motor_task);
        while (true) {
              pros::delay(5);
         }
}
```
The above code ran the motor when opcontrol was started and stopped the motor when the robot was disabled. This enabled/disabled switch was repeated multiple times and the motor still worked as expected.

[x] Run the above test code globally. Got the same results, the motor runs when it is supposed to in opcontrol.

Reviewers: baylessj, O1 The Ori, brookea

Reviewed By: baylessj, O1 The Ori, brookea

Subscribers: brookea

Tags: #zorp

Differential Revision: https://phabricator.purduesigbots.com/D222