Prepare for open source
Summary:
- Add libv5rts as submodule
- Update makefile to point to new SDK locations
- Remove references to include/ifi
  - in headers
  - in arclint

(AKA I'm amazed this worked the first time)

I'm sure we'll probably get some angry friends when they find out how we've gated the SDK, but it is what it is. The SDK is proprietary, and we were not given permission to distribute it. As far as community contributions go, they should be able to extract symbols from libpros.a, and with access to the PROS source, they'll be able to see what functions and types need to be declared.

Test Plan:
- [X] Make sure there are no issues with symbols stripped by `fix_libv5rts`
The above was tested by compiling with a call to `printf` in opcontrol (after the `fix-libv5rts` rule was added in a bunch of places). There were no linker errors.

Run the following:
- [X] `git submodule init` to initialize the .gitmodules file
- [X] `git submodule update` to clone the submodule
- [X] `prosv5 build-compile-commands all` to make sure it all works

Optional(?)
- [ ] Test above steps with an account that does not have permissions to clone purduesigbots/libv5rts (one of the first two should fail)

Reviewers: O1 The Ori, berman5

Reviewed By: O1 The Ori, berman5

Subscribers: berman5

Tags: #zorp

Differential Revision: https://phabricator.purduesigbots.com/D265