🔥 ⚠ Shuffle Makefiles (#87)
#### Summary:
- Move most targets and variables into common.mk. The only thing that remains in Makefile and template-Makefile are "parameters" to common.mk
- Projects that are libraries (setups like PROS kernel and okapilib) will compile the library and link a final uploadable product against that. Rather than linking directly against all the `.o` files that the library archives, we link against libpros.a and any files not included in libpros.a inside `src/`

#### Motivation:
Files in projects marked as "user files" by the template are never modified - when upgrading a project, only "system files" are allowed to be modified. We'd like to be able to modify how PROS projects compile from version to version. Currently, a lot of the logic for compiling projects lives in Makefile. This PR moves as much logic as possible into the system file common.mk.

I'm particularly motivated to get this done so that we can make modifications for hot-cold linking.

#### Impact:
User projects must manually update their Makefile to remove everything we're removing in this PR.

##### References (optional):
Supporting T741

#### Test Plan:
- [X] Compile kernel as a complete program
   - [X] Upload default program and runs correctly (sufficient to test libv5rts, kernel, and libc link correctly)
- [X] Compile libpros.a
- [X] Compile template
- [X] Compile user program
   - [X] Upload default program and runs correctly (sufficient to test libv5rts, kernel, and libc link correctly)