✨screen-return-val-fix
Summary:
Adding in error handling for the methods of the screen class. Change all relevant return methods & documentation to accurately reflect changes.

Motivation:
As most of the methods don't have proper error handling, if they were to error it could cause confusion as to where the error stemmed from. This feature is meant to improve clarity.