🚸Implement Error Return Values For Screen API (#344)
* ✨screen-return-val-fix

Summary:
Adding in error handling for the methods of the screen class

Motivation:
As most of the methods don't have proper error handling, if they were to error it could cause confusion as to where the error stemmed from. This feature is meant to improve clarity.

* ✨screen-return-val-fix

Summary:
Adding in error handling for the methods of the screen class. Change all relevant return methods & documentation to accurately reflect changes.

Motivation:
As most of the methods don't have proper error handling, if they were to error it could cause confusion as to where the error stemmed from. This feature is meant to improve clarity.

* Fixed various typos

A few typos & mismatching types were made, all of which were hopefully corrected

* Made suggested edits/corrections

Kept descriptions from extending past column 80, renamed MUTEX_ERROR to E_TOUCH_ERROR, and removed error handling in the touch_handle_task daemon

* Added semicolons

>:(

* Formatting

Added slash and arrow